Tidy (iCalendarFileExportPrinter), refs 338
